requirements.txt
flask==3.0.2
openai==1.12.0
python-dotenv==1.0.1
feedparser==6.0.11
requests==2.31.0
pandas==2.2.0
gunicorn==21.2.0 